lazy load react native webpack conf
